test(channels): split ingress watchdog handoff proof
Keep ingress-monitor.test.ts under the test max-lines budget.
