+++ src/channels/message/ingress-monitor.test.ts
@@ -16,7 +16,6 @@ import {
   type ChannelIngressMonitorLifecycle,
   type CreateChannelIngressMonitorOptions,
 } from "./ingress-monitor.js";
-import { markIngressBoundedProcessingStarted } from "./ingress-processing-handoff.js";
 import { createChannelIngressQueue, type ChannelIngressQueue } from "./ingress-queue.js";
 import {
   ChannelIngressUnavailableError,
@@ -268,83 +267,6 @@ describe("channel ingress monitor", () => {
         queue.enqueue("event-finalizing", { version: 1, rawEvent: "duplicate" }),
       ).resolves.toMatchObject({ kind: "completed" });
       await monitor.stop();
-    });
-  });
-
-  it("retries when delivery never enters bounded processing and the ingress watchdog fires", async () => {
-    vi.useFakeTimers();
-    await withQueue(async (queue) => {
-      let releaseDelivery!: () => void;
-      const held = new Promise<void>((resolve) => {
-        releaseDelivery = resolve;
-      });
-      let adopted = 0;
-      const monitor = createMonitor(queue, async (_raw, lifecycle) => {
-        await held;
-        await lifecycle.onAdopted();
-        adopted += 1;
-      });
-      try {
-        monitor.start();
-        await monitor.admit({ id: "event-watchdog-retry", lane: "a", text: "hello" });
-        await vi.advanceTimersByTimeAsync(20);
-        expect(await queue.listClaims()).toHaveLength(1);
-        await vi.advanceTimersByTimeAsync(5_000);
-        expect(await queue.listFailed?.({ limit: "all" })).toEqual([]);
-        expect(await queue.listPending({ limit: "all" })).toMatchObject([
-          {
-            id: "event-watchdog-retry",
-            attempts: 1,
-            lastError: expect.stringContaining("handler-timeout"),
-          },
-        ]);
-        expect(adopted).toBe(0);
-      } finally {
-        releaseDelivery();
-        await monitor.stop();
-        vi.useRealTimers();
-      }
-    });
-  });
-
-  it("retains one claim through bounded processing longer than the ingress watchdog", async () => {
-    vi.useFakeTimers();
-    await withQueue(async (queue) => {
-      let releaseDelivery!: () => void;
-      const held = new Promise<void>((resolve) => {
-        releaseDelivery = resolve;
-      });
-      let adopted = 0;
-      const monitor = createMonitor(queue, async (_raw, lifecycle) => {
-        markIngressBoundedProcessingStarted(lifecycle.abortSignal);
-        await held;
-        await lifecycle.onAdopted();
-        adopted += 1;
-      });
-      try {
-        monitor.start();
-        await monitor.admit({ id: "event-processing-hold", lane: "a", text: "hello" });
-        await vi.advanceTimersByTimeAsync(20);
-        expect(await queue.listClaims()).toHaveLength(1);
-        await vi.advanceTimersByTimeAsync(15_000);
-        expect(await queue.listClaims()).toHaveLength(1);
-        expect(await queue.listPending({ limit: "all" })).toEqual([]);
-        expect(await queue.listFailed?.({ limit: "all" })).toEqual([]);
-
-        releaseDelivery();
-        await vi.advanceTimersByTimeAsync(0);
-        await monitor.waitForIdle();
-        expect(adopted).toBe(1);
-        expect(await queue.listClaims()).toEqual([]);
-        expect(await queue.listPending({ limit: "all" })).toEqual([]);
-        expect(await queue.listFailed?.({ limit: "all" })).toEqual([]);
-        await expect(
-          queue.enqueue("event-processing-hold", { version: 1, rawEvent: "duplicate" }),
-        ).resolves.toMatchObject({ kind: "completed" });
-      } finally {
-        await monitor.stop();
-        vi.useRealTimers();
-      }
     });
   });
